+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## et-orbi 1.0.10  not yet released
 
+- Implement EoTime .utc and .local (based on Time .utc and .local)
 - Add EoTime#translate(target_zone) as #localtime(target_zone) alias
 - Correct EoTime#iso8601 (was always returning zulu iso8601 string)